Final commit

+++ events/voiceChannelSystem.js
@@ -1,7 +1,9 @@
 const { Collection, MessageEmbed } = require('discord.js');
+const ms = require('ms');
 const GeneratorSchema = require('../models/generator-schema');
 const voiceCollection = new Collection();
 const chatCollection = new Collection();
+const cooldown = new Collection();
 const client = require('../index');
 
 client.on('voiceStateUpdate', async (oldState, newState) => {
@@ -11,16 +13,42 @@ client.on('voiceStateUpdate', async (oldState, newState) => {
     let generatorID = '';
     let categoryID = '';
 
-    await GeneratorSchema.findOne({ Guild: newState.guild.id }, (err, data) => {
+    await GeneratorSchema.findOne({ Guild: newState.guild.id }, async (err, data) => {
         if (data) {
-            if (newState.guild.channels.cache.get(data.Channel)) generatorID = data.Channel;
-            else return;
-            if (newState.guild.channels.cache.get(data.Category)) categoryID = data.Category;
-            else return console.log('NO PRIVATE VC CATEGORY FOUND!');
+
+            if (newState.guild.channels.cache.find(c => c.id == data.Channel)) generatorID = data.Channel;
+            else if (newState.guild.channels.cache.find(c => c.id == data.Channel) == undefined) {
+                data.Channel = '';
+                data.save();
+                return;
+            }
+            if (newState.guild.channels.cache.find(c => c.id == data.Category)) categoryID = data.Category;
+            else if (newState.guild.channels.cache.find(c => c.id == data.Category) == undefined) {
+                data.Channel = '';
+                data.save();
+                return;
+            }
         }
     })
 
     if (!oldState.channel && newState.channel.id === generatorID) {
+
+        if (cooldown) {
+            if (cooldown.has(`${newState.member.user.id}${newState.guild.id}`))
+                return newState.member.send(
+                    new MessageEmbed()
+                        .setAuthor(newState.member.user.username, newState.member.user.displayAvatarURL({ dynamic: true }))
+                        .setDescription(`Вы не можете создать новый канал в течение \`${ms(cooldown.get(`${newState.member.user.id}${newState.guild.id}`) - Date.now(), { long: false }) }\``)
+                        .setColor('F93A2F')
+                        .setTimestamp()
+                )
+            
+            cooldown.set(`${newState.member.user.id}${newState.guild.id}`, Date.now() + 20000);
+            setTimeout(() => {
+                cooldown.delete(`${newState.member.user.id}${newState.guild.id}`);
+            }, 20000);
+        }
+
         //! CREATE VOICE CHAT
         const vc = await newState.guild.channels.create(`${user.username}'s VC`, {
             type: 'voice',
@@ -57,40 +85,61 @@ client.on('voiceStateUpdate', async (oldState, newState) => {
         });
         chatCollection.set(user.id, chat.id);
 
-        const embed = new MessageEmbed()
-            .setDescription('👻 скрыть канал на сервере и сделать его невидимым\n\n👁 сделать канал снова видимым\n\n🔒 заблокировать канал от входа других пользователей\n\n🔓 разблокировать канал для входа других пользователей\n\n1️⃣ - 5️⃣ ограничение на количество пользователей')
+        const embed1 = new MessageEmbed()
+            .setTitle('Основные настройки')
+            .setDescription('👻 сделать канал невидимым на сервере\n\n👁 сделать канал видимым\n\n🔒 заблокировать вход для других пользователей\n\n🔓 разблокировать вход для других пользователей')
+            .setColor('00D166')
+        const embed2 = new MessageEmbed()
+            .setTitle('Ограничение на количество пользователей')
+            .setDescription('0️⃣ убрать ограничение на количество пользователей\n\n1️⃣ - 5️⃣ ограничение на количество пользователей')
+            .setColor('00D166')
+        const embed3 = new MessageEmbed()
+            .setTitle('Настройки качества звука')
+            .setDescription('🔴 низкое качество звука\n\n🟠 среднее качество звука\n\n🟢 высокое качество звука')
             .setColor('00D166')
 
-        chat.send(`<@${user.id}>`, embed)
+        chat.send(`<@${user.id}>`, embed1)
             .then((embed) => {
                 embed.react('👻'); // make invisible
                 embed.react('👁'); // make visible
-
                 embed.react('🔒'); // do not allow users to join
                 embed.react('🔓'); // allows users to join
-
+            })
+        chat.send(embed2)
+            .then((embed) => {
+                embed.react('0️⃣'); // remove user limit
                 embed.react('1️⃣'); // user limit: 1
                 embed.react('2️⃣'); // user limit: 2
                 embed.react('3️⃣'); // user limit: 3
                 embed.react('4️⃣'); // user limit: 4
                 embed.react('5️⃣'); // user limit: 5
             })
+        
+        chat.send(embed3)
+            .then((embed) => {
+                embed.react('🔴'); // sound quality: low
+                embed.react('🟠'); // sound quality: medium
+                embed.react('🟢'); // sound quality: high
+            })
 
-    } else if (!newState.channel) {
+    } else if (newState.channel ? newState.channel.id : '' != voiceCollection.get(user.id)) {
+
+        if (!oldState.channel) return;
 
         if (oldState.channel.id === voiceCollection.get(newState.id)) {
             oldState.channel.delete(); // delete voice chat
 
             const embed = new MessageEmbed()
-                .setDescription('⚠ Канал будет удалён через 10 секунд!')
+                .setDescription(`⚠ Канал будет удалён через \`12с\``)
                 .setColor('F93A2F')
             newState.guild.channels.cache.get(chatCollection.get(newState.id)).send(embed);
 
-            setTimeout(function() { deleteTextChannel(newState) }, 9000) // delete text chat (after 9 seconds)
+            setTimeout(function() { deleteTextChannel(newState) }, 12 * 1000) // delete text chat (after 12 seconds)
             return;
         }
 
         function deleteTextChannel(chat) {
+            if (!chat.guild.channels.cache.get(chatCollection.get(chat.id))) return;
             chat.guild.channels.cache.get(chatCollection.get(chat.id)).delete() 
         }
     }
@@ -153,4 +202,28 @@ function limit5(messageReaction, user) {
     if (vc) vc.setUserLimit(5);
 }
 
-module.exports = { makeInvisible, makeVisible, makeJoinable, makeNotJoinable, limit1, limit2, limit3, limit4, limit5 };
+//* SETS USER LIMIT TO 5
+function removeLimit(messageReaction, user) {
+    const vc = messageReaction.message.guild.channels.cache.get(voiceCollection.get(user.id));
+    if (vc) vc.setUserLimit(0);
+}
+
+//* SETS LOW BITRATE
+function bitrateLow(messageReaction, user) {
+    const vc = messageReaction.message.guild.channels.cache.get(voiceCollection.get(user.id));
+    if (vc) vc.edit({ bitrate: 8000 });
+}
+
+//* SETS MIDDLE BITRATE
+function bitrateMiddle(messageReaction, user) {
+    const vc = messageReaction.message.guild.channels.cache.get(voiceCollection.get(user.id));
+    if (vc) vc.edit({ bitrate: 64000 });
+}
+
+//* SETS HIGH BITRATE
+function bitrateHigh(messageReaction, user) {
+    const vc = messageReaction.message.guild.channels.cache.get(voiceCollection.get(user.id));
+    if (vc) vc.edit({ bitrate: 96000 });
+}
+
+module.exports = { makeInvisible, makeVisible, makeJoinable, makeNotJoinable, limit1, limit2, limit3, limit4, limit5, removeLimit, bitrateLow, bitrateMiddle, bitrateHigh };
